# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lwang Black API",
    "description": "E-commerce backend API for Lwang Black specialty coffee. Supports multi-currency payments (Stripe, PayPal, Khalti, eSewa, Nabil Bank), global logistics (Shippo, DHL, FedEx), admin dashboard, and notification services.",
    "version": "1.0.0",
    "contact": {
      "email": "brewed@lwangblack.co",
      "url": "https://www.lwangblack.co"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" },
    { "url": "https://api.lwangblack.co", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/notifications/log": {
      "get": {
        "tags": ["Notifications"],
        "summary": "View notification history (email/SMS)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Notification log entries" } }
      }
    },
    "/api/notifications/send-email": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Send manual email (SendGrid)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Email sent" } }
      }
    },
    "/api/notifications/send-sms": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Send manual SMS (Twilio)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "SMS sent" } }
      }
    },
    "/api/notifications/invoice/{orderId}": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Generate and optionally email invoice PDF",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Invoice generated" } }
      },
      "get": {
        "tags": ["Notifications"],
        "summary": "Get invoice for an order",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Invoice details" } }
      }
    },
...
  }
}